Type-conversion warning fixes

+++ src/gui/midi.cpp
@@ -509,7 +509,7 @@ void MIDI_State_LoadMessage()
 void MIDI_RawOutByte(Bit8u data) {
 	if (midi.sysex.start) {
 		Bit32u passed_ticks = GetTicks() - midi.sysex.start;
-		if (passed_ticks < midi.sysex.delay) SDL_Delay(midi.sysex.delay - passed_ticks);
+		if (passed_ticks < midi.sysex.delay) SDL_Delay((Uint32)(midi.sysex.delay - passed_ticks));
 	}
 
 	/* Test for a realtime MIDI message */
